Revert "Security setups"
