--- src/Data/DAL/Security/IPermissionRepository.cs
@@ -1,13 +1,0 @@
-﻿using Domain.Entity.Security;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Data.DAL.Security
-{
-   public interface IPermissionRepository : IRepository<Guid, Permission>
-   {
-   }
-}
