--- src/Domain/Entity/Security/User.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Domain.Entity.Security
-{
-   public class User : Item
-   {
-      public User()
-      {
-         Roles = new List<Role>();
-         Name = new PersonName();
-      }
-
-      public virtual string Login { get; set; }
-      public virtual string PasswordHash { get; set; }
-      public virtual PersonName Name { get; set; }
-      public virtual DateTime? PasswordExpires { get; set; }
-      public virtual IList<Role> Roles { get; set; }
-   }
-}
