+++ src/Domain/Entity/Setup/Role.cs
@@ -1,0 +1,18 @@
+﻿using System.Collections.Generic;
+
+namespace Domain.Entity.Setup
+{
+    public class Role
+    {
+        public Role()
+        {
+            this.Users = new List<User>();
+        }
+
+        public virtual int RoleID { get; set; }
+        public virtual string Name { get; set; }
+        public virtual string Description { get; set; }
+
+        public virtual ICollection<User> Users { get; set; }
+    }
+}
