+++ src/PrizmMainProject/Forms/Settings/SettingsXtraForm.cs
@@ -10,6 +10,7 @@ using Ninject.Parameters;
 
 using Domain.Entity.Setup;
 
+using PrizmMain.DummyData;
 using PrizmMain.Forms.Settings.Dictionary;
 using PrizmMain.Forms.Settings.UserRole.Role;
 using PrizmMain.Forms.Settings.UserRole.User;
@@ -21,7 +22,6 @@ using Domain.Entity;
 using PrizmMain.Forms.Settings.ViewTypes;
 using PrizmMain.Common;
 using DevExpress.XtraLayout.Customization;
-using Domain.Entity.Security;
 using Domain.Entity.Mill;
 
 namespace PrizmMain.Forms.Settings
@@ -92,12 +92,10 @@ namespace PrizmMain.Forms.Settings
 
             gridViewWelders.BestFitColumns();
             gridViewInspectors.BestFitColumns();
-
-            RefreshUserRoles(0);
-            RefreshRolePermissions(0);
         }
 
-        private void BindToViewModel()
+
+       private void BindToViewModel()
         {
             #region Data Source
             pipeMillSizeTypeBindingSource.DataSource = viewModel;
@@ -120,16 +118,6 @@ namespace PrizmMain.Forms.Settings
 
             repositoryItemsСategory.DataSource = viewModel.CategoryTypes;
             categoriesGrid.DataSource = viewModel.CategoryTypes;
-
-            rolesBindingSource.DataSource = viewModel.Roles;
-            rolesBindingSource.ListChanged += (s, e) => IsModified = true;
-
-            permissionsBindingSource.DataSource = viewModel.Permissions;
-
-            usersBindingSource.DataSource = viewModel.Users;
-
-            gridControlRoles.DataSource = rolesBindingSource;
-
             #endregion
 
             #region Data Bindings
@@ -334,200 +322,10 @@ namespace PrizmMain.Forms.Settings
         {
         }
 
-        private void gridViewRole_InitNewRow(object sender, InitNewRowEventArgs e)
-        {
-           var view = sender as GridView;
 
-           if (view.IsValidRowHandle(e.RowHandle))
-           {
-              Role role = view.GetRow(e.RowHandle) as Role;
-              
-           }
-        }
 
-        private void gridViewRole_ValidateRow(object sender, ValidateRowEventArgs e)
-        {
-           var view = sender as GridView;
-           view.ClearColumnErrors();
-           Role role = e.Row as Role;
-           if (role != null)
-           {
-              if (String.IsNullOrEmpty(role.Name))
-              {
-                 e.Valid = false;
-                 view.SetColumnError(colRoleSetupName, Resources.VALUE_REQUIRED);
-              }
-           }
-        }
 
-        private void gridViewRole_InvalidRowException(object sender, InvalidRowExceptionEventArgs e)
-        {
-           e.ExceptionMode = DevExpress.XtraEditors.Controls.ExceptionMode.NoAction;
-        }
 
-        void RefreshRolePermissions(int rowIndex)
-        {
-           var view = gridViewRole;
-
-           if (view.IsValidRowHandle(rowIndex))
-           {
-              gridViewPermissions.ClearSelection();
-              var role = view.GetRow(rowIndex) as Role;
-              if (role != null)
-              {
-                 for (int rowHandle = 0; rowHandle < gridViewPermissions.RowCount; rowHandle++)
-                 {
-                    var perm = gridViewPermissions.GetRow(rowHandle) as Permission;
-                    if (viewModel.RoleHasPermission(role, perm))
-                    {
-                       gridViewPermissions.SelectRow(rowHandle);
-                    }
-                 }
-              }
-           }
-        }
-
-        private void gridViewRole_FocusedRowChanged(object sender, FocusedRowChangedEventArgs e)
-        {
-           RefreshRolePermissions(e.FocusedRowHandle);
-        }
-
-        private void gridViewPermissions_SelectionChanged(object sender, DevExpress.Data.SelectionChangedEventArgs e)
-        {
-           var view = sender as GridView;
-           var role = gridViewRole.GetFocusedRow() as Role;
-
-           if (role == null)
-              return;
-
-           Permission p = view.GetRow(e.ControllerRow) as Permission;
-
-           switch(e.Action)
-           {
-              case CollectionChangeAction.Add:
-                 viewModel.AddPermissionToRole(role, p);
-                 IsModified = true;
-                 break;
-              case CollectionChangeAction.Remove:
-                 viewModel.RemovePermissionFromRole(role, p);
-                 IsModified = true;
-                 break;
-           }
-        }
-
-        private void gridViewUsers_ValidateRow(object sender, ValidateRowEventArgs e)
-        {
-           var view = sender as GridView;
-
-           if (view.IsValidRowHandle(e.RowHandle))
-           {
-              view.ClearColumnErrors();
-              User user = view.GetRow(e.RowHandle) as User;
-              if (String.IsNullOrEmpty(user.Login))
-              {
-                 view.SetColumnError(colLogin, Resources.VALUE_REQUIRED);
-                 e.Valid = false;
-                 return;
-              }
-              if (String.IsNullOrEmpty(user.PasswordHash))
-              {
-                 view.SetColumnError(colUserPass, Resources.VALUE_REQUIRED);
-                 e.Valid = false;
-                 return;
-              }
-           }
-        }
-
-        private void gridViewUsers_InvalidRowException(object sender, InvalidRowExceptionEventArgs e)
-        {
-           e.ExceptionMode = DevExpress.XtraEditors.Controls.ExceptionMode.NoAction;
-        }
-
-        private void gridViewUsers_InitNewRow(object sender, InitNewRowEventArgs e)
-        {
-           var view = sender as GridView;
-           if (view.IsValidRowHandle(e.RowHandle))
-           {
-              User user = view.GetRow(e.RowHandle) as User;
-              if (user != null)
-              {
-                 user.IsActive = true;
-              }
-           }
-        }
-
-        private void repositoryItemButtonUserPass_ButtonClick(object sender, DevExpress.XtraEditors.Controls.ButtonPressedEventArgs e)
-        {
-           var view = gridViewUsers;
-           if (view.IsValidRowHandle(view.FocusedRowHandle))
-           {
-              User user = view.GetRow(view.FocusedRowHandle) as User;
-              PasswordChangeDialog dlg = new PasswordChangeDialog();
-              if (dlg.ShowPasswordDialog(user.PasswordHash) == System.Windows.Forms.DialogResult.OK)
-              {
-                 user.PasswordHash = dlg.NewPasswordHash;
-                 IsModified = true;
-              }
-           }
-        }
-
-        private void gridViewUsers_KeyDown(object sender, System.Windows.Forms.KeyEventArgs e)
-        {
-           var view = sender as GridView;
-           view.RemoveSelectedItem(e, viewModel.Users, (_) => _.IsNew());
-        }
-
-        private void gridViewRoles_SelectionChanged(object sender, DevExpress.Data.SelectionChangedEventArgs e)
-        {
-           var view = sender as GridView;
-           var user = gridViewUsers.GetRow(gridViewUsers.FocusedRowHandle) as User;
-
-           if (user != null)
-           {
-              var role = view.GetRow(e.ControllerRow) as Role;
-              if (role != null)
-              {
-                 switch (e.Action)
-                 {
-                    case CollectionChangeAction.Add:
-                       viewModel.AddRoleToUser(role, user);
-                       IsModified = true;
-                       break;
-                    case CollectionChangeAction.Remove:
-                       viewModel.RemoveRoleFromUser(role, user);
-                       IsModified = true;
-                       break;
-                 }
-              }
-           }
-        }
-
-        void RefreshUserRoles(int currentRow)
-        {
-           var view = gridViewUsers;
-
-           if (view.IsValidRowHandle(currentRow))
-           {
-              gridViewRoles.ClearSelection();
-              var user = view.GetRow(currentRow) as User;
-              if (user != null)
-              {
-                 for (int rowHandle = 0; rowHandle < gridViewRoles.RowCount; rowHandle++)
-                 {
-                    var role = gridViewRoles.GetRow(rowHandle) as Role;
-                    if (role != null && viewModel.UserHasRole(user, role))
-                    {
-                       gridViewRoles.SelectRow(rowHandle);
-                    }
-                 }
-              }
-           }
-        }
-
-        private void gridViewUsers_FocusedRowChanged(object sender, FocusedRowChangedEventArgs e)
-        {
-           RefreshUserRoles(e.FocusedRowHandle);
-        }
 
     }
