+++ src/PrizmMainProject/Security/ISecurityContext.cs
@@ -1,5 +1,4 @@
-﻿using Domain.Entity.Security;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -10,6 +9,5 @@ namespace PrizmMain.Security
     {
         bool HasAccess(Privileges privilege);
         Domain.Entity.PersonName GetLoggedPerson();
-        User LoggedUser { get; set; }
     }
 }
