+++ src/PrizmMainProject/Security/Privileges.cs
@@ -8,7 +8,6 @@ namespace PrizmMain.Security
 {
     public enum Privileges
     {
-       ReadSettings,
-       WriteSettings
+
     }
 }
