+++ src/UnitTests/Forms/Railcar/Edit/SaveRailcarCommandTest.cs
@@ -2,7 +2,6 @@
 using Domain.Entity.Mill;
 using Moq;
 using NUnit.Framework;
-using PrizmMain.Documents;
 using PrizmMain.Forms;
 using PrizmMain.Forms.Railcar;
 using PrizmMain.Forms.Railcar.NewEdit;
@@ -25,7 +24,6 @@ namespace UnitTests.Forms.Railcar.Edit
             repos.SetupGet(_ => _.RailcarRepo).Returns(railcarRepo.Object);
 
             var viewModel = new RailcarViewModel(repos.Object, "", notify.Object);
-            viewModel.ModifiableView = new Mock<IModifiable>().Object;
             viewModel.Railcar.Number = "Test Railcar";
             viewModel.Railcar.Pipes.Add(new Pipe());
             var command = new SaveRailcarCommand(viewModel, repos.Object, notify.Object);
