+++ src/UnitTests/Forms/Settings/SaveSettingsCommandTest.cs
@@ -13,8 +13,6 @@ using System.Threading.Tasks;
 using Domain.Entity;
 using Domain.Entity.Mill;
 using PrizmMain.Forms;
-using Data.DAL.Security;
-using Domain.Entity.Security;
 using PrizmMain.Documents;
 
 
@@ -28,15 +26,15 @@ namespace UnitTests.Forms.Settings
         {
             var notify = new Mock<IUserNotify>();
 
+            var modifiableView = new Mock<IModifiable>();
+
             var repoPipeSize = new Mock<IMillPipeSizeTypeRepository>();
             var repoPipeTests = new Mock<IPipeTestRepository>();
             var repoWelders = new Mock<IWelderRepository>();
             var repoInspectors = new Mock<IInspectorRepository>();
             var repoManufacturers = new Mock<IPlateManufacturerRepository>();
             var repoProjectSetting = new Mock<IProjectRepository>();
-            var repoUsers = new Mock<IUserRepository>();
-            var repoRoles = new Mock<IRoleRepository>();
-            var repoPerms = new Mock<IPermissionRepository>();
+
             var repoCategory = new Mock<ICategoryRepository>();
             
             var testProjectSetting = new Project();
@@ -44,20 +42,13 @@ namespace UnitTests.Forms.Settings
             var testWelder = new Welder();
             var testManufacturer = new PlateManufacturer();
             var testInspector = new Inspector();
-            var testUser = new User();
-            var testRole = new Role();
-            var testPerm = new Permission();
-            var modifiableView = new Mock<IModifiable>();
 
             repoPipeSize.Setup(_ => _.GetAll()).Returns(new List<PipeMillSizeType>() { testSizeType });
             repoWelders.Setup(_ => _.GetAll()).Returns(new List<Welder>() { testWelder });
             repoManufacturers.Setup(_ => _.GetAll()).Returns(new List<PlateManufacturer>() { testManufacturer });
             repoProjectSetting.Setup(_ => _.GetSingle()).Returns(testProjectSetting);
             repoInspectors.Setup(_ => _.GetAll()).Returns(new List<Inspector>() { testInspector });
-            repoUsers.Setup(_ => _.GetAll()).Returns(new List<User>() { testUser });
-            repoRoles.Setup(_ => _.GetAll()).Returns(new List<Role>() { testRole });
-            repoPerms.Setup(_ => _.GetAll()).Returns(new List<Permission>() { testPerm });
-            
+
             repoCategory.Setup(x => x.GetAll()).Returns(new List<Category>() { new Category() });
 
             Mock<ISettingsRepositories> settingsRepos = new Mock<ISettingsRepositories>();
@@ -67,13 +58,13 @@ namespace UnitTests.Forms.Settings
             settingsRepos.SetupGet(_ => _.PlateManufacturerRepo).Returns(repoManufacturers.Object);
             settingsRepos.SetupGet(_ => _.ProjectRepo).Returns(repoProjectSetting.Object);
             settingsRepos.SetupGet(_ => _.InspectorRepo).Returns(repoInspectors.Object);
-            settingsRepos.SetupGet(_ => _.UserRepo).Returns(repoUsers.Object);
-            settingsRepos.SetupGet(_ => _.RoleRepo).Returns(repoRoles.Object);
-            settingsRepos.SetupGet(_ => _.PermissionRepo).Returns(repoPerms.Object);
             settingsRepos.SetupGet(x => x.СategoryRepo).Returns(repoCategory.Object);
 
+            modifiableView.SetupGet(x => x.IsModified).Returns(false);
+
+
             var viewModel = new SettingsViewModel(settingsRepos.Object, notify.Object);
-            viewModel.ModifiableView = modifiableView.Object;
+
             viewModel.LoadData();
             viewModel.ModifiableView = modifiableView.Object;
 
@@ -93,10 +84,6 @@ namespace UnitTests.Forms.Settings
             repoProjectSetting.Verify(_ => _.Evict(testProjectSetting), Times.Once());
             repoInspectors.Verify(_ => _.SaveOrUpdate(testInspector), Times.Once());
             repoInspectors.Verify(_ => _.Evict(testInspector), Times.Once());
-            repoUsers.Verify(_ => _.SaveOrUpdate(testUser), Times.Once);
-            repoUsers.Verify(_ => _.Evict(testUser), Times.Once);
-            repoRoles.Verify(_ => _.SaveOrUpdate(testRole), Times.Once);
-            repoRoles.Verify(_ => _.Evict(testRole), Times.Once);
         }
     }
 }
