+++ src/UnitTests/UnitTests.csproj
@@ -62,7 +62,6 @@
   <ItemGroup>
     <Compile Include="DisplayObjectInfo.cs" />
     <Compile Include="AuditTrailTests.cs" />
-    <Compile Include="Forms\Common\PasswordEncryptorTest.cs" />
     <Compile Include="Forms\PrizmMainForm.cs" />
     <Compile Include="Forms\PipeMill\NewEdit\MillPipeNewEditCommandTest.cs" />
     <Compile Include="Forms\PipeMill\Search\MillPipeSearchCommandTest.cs" />
